Printing Multiple Lines
Write the Java code to display the given lines as it is on screen using the knowledge of print statements taught in the above lessons.             In the realm of atoms and galaxies,
Where mysteries unfurl like cosmic seas,
Science, our guiding star so bright,
Reveals the universe, its boundless light.

From microcosms to celestial grace,
Inquiring minds explore every space,
In the quest for knowledge, we stand tall,
Science, the greatest adventure of all.